+++ src/fapilog/_internal/background_cleanup_manager.py
@@ -9,7 +9,7 @@ import logging
 import time
 from typing import Any, Callable, Optional, Protocol
 
-from .async_task_manager import AsyncTaskManager
+from ..async_components.concurrency.task_manager import AsyncTaskManager
 
 logger = logging.getLogger(__name__)
