+++ src/fapilog/_internal/deduplication_processor.py
@@ -10,9 +10,9 @@ import logging
 import time
 from typing import Any, Dict, List, Optional, cast
 
+from ..async_components.concurrency.task_manager import AsyncTaskManager
 from ..exceptions import ProcessorConfigurationError
 from .async_processor_base import AsyncProcessorBase
-from .async_task_manager import AsyncTaskManager
 from .background_cleanup_manager import BackgroundCleanupManager, CleanupTarget
 
 logger = logging.getLogger(__name__)
