+++ tests/test_container_component_registry.py
@@ -10,10 +10,10 @@ from unittest.mock import patch
 
 import pytest
 
-from fapilog._internal.async_lock_manager import ProcessorLockManager
 from fapilog._internal.component_registry import ComponentRegistry
 from fapilog._internal.metrics import MetricsCollector
 from fapilog._internal.processor_metrics import ProcessorMetrics
+from fapilog.async_components.concurrency.lock_manager import ProcessorLockManager
 from fapilog.container import LoggingContainer
 from fapilog.core.factories.component_factory import ComponentFactory
 from fapilog.monitoring import PrometheusExporter
